move AppQuerier code into GraphQLController
