+++ EcoSoapBank/Network Controller/GraphQLController.swift
@@ -16,9 +16,14 @@ enum HTTPMethod: String {
 class GraphQLController {
 
     // MARK: - Properties
+    
+    var loggedIn: Bool {
+        token != nil
+    }
 
     private let session: DataLoader
     private let url = URL(string: "http://35.208.9.187:9094/ios-api-1/")!
+    private var token: String?
 
     // Setting up the url request
     private lazy var request: URLRequest = {
@@ -63,6 +68,10 @@ class GraphQLController {
 
             completion(self.decodeJSON(type, data: data))
         }
+    }
+    
+    func provideToken(_ token: String) {
+        self.token = token
     }
 
     // MARK: - Helper Methods
@@ -111,4 +120,6 @@ class GraphQLController {
 
 enum GraphQLError: Error {
     case noData
+    case noToken
+    case unimplemented
 }
